python_sim/requirements.txt
# Core numerics / data
pandas>=2.2.2
numpy>=2.0.2
scipy>=1.12.0

# Geospatial + network building
osmnx>=1.9.3
networkx>=3.2.1
geopandas>=1.0.0
shapely>=2.0.2
pyproj>=3.5.0
pyogrio>=0.9.0
geopy>=2.4.1

# Visualization
matplotlib>=3.9.2
seaborn>=0.13.2

# SUMO TraCI bindings
traci>=1.20.0
sumolib>=1.20.0

# API interactions
requests>=2.32.3
python-dotenv>=1.0.1
typer>=0.12.5

# Experiment management
tqdm>=4.66.5
pyyaml>=6.0.2

